fix: address code review — empty array dtype, missing test coverage

+++ gpt2_triton/kernels/gelu.py
@@ -60,7 +60,7 @@ def gelu(x: np.ndarray) -> np.ndarray:
     y : np.ndarray, same shape as ``x``.
     """
     if x.size == 0:
-        return np.empty_like(x)
+        return np.empty_like(x, dtype=np.float32)
 
     x_dev = gpu.to_device(np.ascontiguousarray(x).astype(np.float32))
     y_dev = gpu.allocate(x.shape, np.float32)
